Enable checkout from a CVS server equipped with CVSACL.
Pierre Poissinger wrote:

git cvsimport fails to checkout from a CVS server patched with CVSACL
(http://cvsacl.sourceforge.net/).
The problem is related to the cvsps requesting version to such server
and expecting only 2 lines... but CVSACL adds a additional line to the
resp. of the version query
==> cvsps shouts a "cvs_direct: protocol error reading version" and go
tits up...

I just added a "special" case for this CVSACL line (maybe cleaner to
actually loop until the "ok" from cvs, but then I was afraid that, in
some conditions, it will result in a dead loop - didn't see any one
getting this specific issue so, better be safe :-) )

Btw, with this patch, my export went fine :-)