Fix branch detection in the basic.repo test.
I'm a little suspicious about this one, though.  It seems too easy.  The
previois version to the line didn't make amy sense - attaching the
pre-revision branch children to the post-revision list???  This was
either a bonehead error or a piece of subtle magic I don't understand.
I hope it was the former...