git-cvsimport: handle the backend option first
Most of the options to git-cvsimport are simply passed to the backend,
this means the if the '-e <backend>' option does not occur first any
options preceding it will effectively be ignored.

Extract the '-e' option handling as a special case before the main
option handling loop to avoid this issue.

Signed-off-by: John Keeping <john@keeping.me.uk>