Make UTC-offset calcuation more portable.
The original implementation of the timezone logic was ported from Git's
cvsimport script for the 2.x versions of cvsps.  A portability issue was
pointed out by Ben Walton on the Git list:

  http://comments.gmane.org/gmane.comp.version-control.git/213686

In summary, the conversion specifications "%s" and "%z" of strftime(3)
are not supported on some platforms (e.g., Solaris 10.)  Junio Hamano
suggested an implementation for the Perl script which only depends on
localtime(3).  Port this idea for cvsps, also using mktime(3) instead of
implementing Perl's timegm() functionality.

Signed-off-by: Chris Rorvick <chris@rorvick.com>
Signed-off-by: Eric S. Raymond <esr@thyrsus.com>